fix compile warning on MSVC

+++ arangod/Cluster/EnsureIndex.cpp
@@ -92,7 +92,7 @@ EnsureIndex::~EnsureIndex() = default;
 
 // For local book keeping and reporting on /_admin/actions
 arangodb::Result EnsureIndex::setProgress(double d) {
-  _progress = d;
+  _progress = static_cast<uint64_t>(d);
   return {};
 }
